+++ src/realm/objc/test/table_delete_all.m
@@ -8,6 +8,7 @@
 #import <realm/objc/RLMTable.h>
 #import <realm/objc/RLMDescriptor.h>
 #import <realm/objc/PrivateRLM.h>
+#import <realm/objc/RLMTableFast.h>
 
 @interface MACTestTableDeleteAll: RLMTestCase
 @end